Develop

+++ app/src/main/java/by/godevelopment/currencyexchangeapplication/domain/models/CurrencyModel.kt
@@ -2,10 +2,11 @@ package by.godevelopment.currencyexchangeapplication.domain.models
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
+import java.math.BigDecimal
 
 data class CurrencyModel(
     val id: Int,
-    val rate : Double,
+    val rate : BigDecimal,
     val base : String,
     @StringRes
     val currencyName: Int,
